Only paint regular video frames via the composited path.
In the software path, always paint the poster image if present, or otherwise nothing.

If GlobalPaintFlattenCompositingLayers is passed when painting videos via the
software path, paint the actual video frame. This is to support printing and
drag-and-drop.

BUG=527579

Review URL: https://codereview.chromium.org/1579423002

Cr-Original-Commit-Position: refs/heads/master@{#369298}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf4e52129f8305d7c1f938da837ba1db80bf201c