Fix word-wrap: break-word with fit-content not to break surrogate pairs
This patch fixes not to break surrogate pairs when word-wrap:
break-word with fit-content.

This combination needs a larger fix in crbug.com/380667, but this fix
is small enough to take at this point, and it is reasonable to handle
trailing pairs when we handle leading pairs.

BUG=574577

Review URL: https://codereview.chromium.org/1575333002

Cr-Original-Commit-Position: refs/heads/master@{#369335}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f417f950dcde37beac951cce67e52025affd22bf