Fix behavior of Document.body getter
According to [1], the body element is the first child of the html
element that is either a body element or a frameset element. The
current version does not check that the parent node is the html
element, so add that check.

Behavior matches Firefox.

[1] https://html.spec.whatwg.org/multipage/dom.html#the-body-element-2

BUG=490511

Review URL: https://codereview.chromium.org/1578193006

Cr-Original-Commit-Position: refs/heads/master@{#369654}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c482a252feb7731369dcc483b0b145efaae7aba6