Fix StylePendingImage objects leaking out of style resolving
This patch fixes a crash caused by serialising StylePendingImages
left on a ComputedStyle's contentData when the image's URL uses a
disallowed protocol.

The fix is to create a new StyleInvalidImage object that is able to
serialise to the original URL string instead of crashing and retain the
contract that StylePendingImages do not live beyond style resolving.

BUG=573790

Review URL: https://codereview.chromium.org/1571393005

Cr-Original-Commit-Position: refs/heads/master@{#369978}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ffb693b5dbc7724ced3d7520658c9e9db5def14