This CL put [Exposed=(Window,Worker)] in the ImageBitmap.idl
Along with that, some of the layout tests results has to be updated.

Also, a layout test has been updated to show that the worker is
able to get width&height for an ImageBitmap.

The thread of Intent to ship ImageBitmap can be found here:
https://groups.google.com/a/chromium.org/d/topic/blink-dev/jRzvehX9U1U/discussion

BUG=563984

Review URL: https://codereview.chromium.org/1486663002

Cr-Original-Commit-Position: refs/heads/master@{#370179}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0b95bd78dc3ac8936f7dec1dce92ec1118d04f95