Enable DumpAccessibilityTree tests to use cross-process iframes.
Makes it possible for a DumpAccessibilityTree test to load an iframe
in a separate process simply by adding the text @CROSS-SITE: to the url,
for example:

  <iframe src="@CROSS-SITE:iframe.html"></iframe>

The test framework handles the details automatically including blocking
until the accessibility tree for the child frame finishes loading.
This significantly reduces the difficulty and complexity of adding new
tests for accessibility of out-of-process iframes.

This change also tries to improve the robustness of existing iframe
accessibility tests and re-enables a previously flaky test.

BUG=368298, 224659, 551601

Review URL: https://codereview.chromium.org/1552683002

Cr-Original-Commit-Position: refs/heads/master@{#370192}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ff8a7383fdcfa996b2c673fdaafbab2eedd05d8c