Fix links in inline continuations in printed PDF
We output links as clickable rects to PDF when printing. If a
link contains continuations, the paint rect intersection check
of the containing anonymous block may cause miss of output of
the PDF rect, because the visual overflow rect of the anonymous
block is empty or is smaller than the link PDF rect.

Now when printing, we let the parent block check for intersection
for the anonymous block of the link.

BUG=535514
TEST=PrintContextTest.LinkTargetComplex

Review URL: https://codereview.chromium.org/1592493002

Cr-Original-Commit-Position: refs/heads/master@{#370260}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 83c02707dd7c8d48c47503d10734d67dd18ce254