Adds a name function for constraints.
The constraints specification requires that one return the
name of a failing constraint in an error message. It's
therefore required to be able to extract the name to return
from the constraint.

This CL adds that function.

BUG=543997

Review URL: https://codereview.chromium.org/1601873002

Cr-Original-Commit-Position: refs/heads/master@{#370378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1bb1840db5ac2d6af5bd73ee2b498bc28867c095