Deprecate fetching stylesheets with unsupported type
The spec http://www.w3.org/TR/html5/document-metadata.html#the-link-element
specifies that "If the UA does not support the given MIME type for the given
link relationship, then the UA should not obtain the resource".

The linked bug contains a testcase that illustrates that Blink is not currently
doing this, and is obtaining the resource even if the link type is e.g.
"application/javascript".

This patch translates the testcase from the linked bug into Blink layout tests
and implements a deprecation warning ahead of implementing the suggested fix.

For more background and discussion, see
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/nH1O6WszMgo/discussion

BUG=286682

Review URL: https://codereview.chromium.org/1501393003

Cr-Original-Commit-Position: refs/heads/master@{#370531}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db6584f5ba2d6cda49fa42775903399cd4d427d6