Implement v1 slot logic in EventPath
This CL makes sure that when an event happens on an element which is assigned to a slot, the event path goes to the slot after the element and goes up the v1 tree.
Described here in spec.
http://w3c.github.io/webcomponents/spec/shadow/#event-paths

BUG=531990

Review URL: https://codereview.chromium.org/1606153002

Cr-Original-Commit-Position: refs/heads/master@{#370648}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b457ca4d9a86e52c11d906b56f55bdf4c6f1136f