Don't force display:table-cell and float:none on table cell elements when in quirks mode
Currently in quirks mode we force table cell elements to
display:table-cell and float:none. This current action is not supported
in quirks spec and is inconsistent with Gecko. This leads to elements
with display/float set to lose that information. This patch fixes that
and allows those elements with display/float to retain this state even
in quirks mode.

BUG=580053

Review URL: https://codereview.chromium.org/1612273002

Cr-Original-Commit-Position: refs/heads/master@{#370904}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95773a85549dc14dbb42d665dadd7875d05df0a1