[css-grid] Add parsing support for <auto-repeat> syntax
The repeat() notation allows now to specify auto-fill or auto-fit instead of
a fixed number of repetitions meaning that it will be automatically computed
depending on the available space.

This CL just adds the parsing support, the expansion of the repeat notation
will be implemented in a follow up patch because it cannot be done at
parsing level (since it requires knowing about the available space).

BUG=579104

Review URL: https://codereview.chromium.org/1582793002

Cr-Original-Commit-Position: refs/heads/master@{#370951}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7647a0d4b25582406b03216f63f4906bd3c46d22