Add a key validator for the experimental framework
This is an initial implementation of the actual API key validator for the
experimental framework. This verifies in content that a key obtained in Blink
is appropriate for use in a given context, and contains a valid signature.

BUG=543215

Review URL: https://codereview.chromium.org/1601753002

Cr-Original-Commit-Position: refs/heads/master@{#370978}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5eb601998ccddb6788476f84df6c402e6387fb7