Add new media image layout tests for video color profiles
  media/color-profile-munsell-bt601-smpte-to-srgb.html
  media/color-profile-munsell-bt709-to-srgb.html

The tests transform a test chart image from source color profile
(bt601, bt709) to sRGB and compares the output with the expected
Munsell colors [1] and studio video range black and white colors
and some proportions of them, SG Black 2x for example. Note: all
the expected colors are in sRGB color space.

The tests (one each for bt601 & bt709) are designed to be run in
the test harness automatically, but also manually to help detect
color changes should that happen in future.

The tabulated color results of these tests are correct and don't
change or need to be rebaselined. If the test colors changed, it
is a colorimetric error: the offending CL should be reverted.

[1] For more details on Munsell colors and gamuts, see
http://www.brucelindbloom.com/index.html?ColorCheckerRGB.html

TBR=kbr@chomium.org
BUG=570150

Review URL: https://codereview.chromium.org/1629963005

Cr-Original-Commit-Position: refs/heads/master@{#371589}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eec72142e952022f7c4891f667768195b8b4be54