Correct a small mistake in ImageBitmap::copyBitmapData
In a previous CL, it was pointed out that the number "4" should be replaced
by info.bytesPerPixel() in two places. In that CL, it was corrected in
one place, but the other one was accidently left over. This CL corrects
the left-over one.

Review URL: https://codereview.chromium.org/1631353002

Cr-Original-Commit-Position: refs/heads/master@{#371593}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1e65bcf347ed812dc717c883517383419795a35