workers: Move where idle-tasks are turned on for v8::Isolate.
Compositor-workers currently do not support idle tasks (and so does not
have idle-task-runner). So idle-tasks should not be turned on for the
compositor workers' isolate. To fix this, move the code that turns on
idle tasks for the isolate into WorkerThread::initializeIsolate(), so
that compositor worker can choose to not do this for now.

BUG=565789, 430155

Review URL: https://codereview.chromium.org/1612093003

Cr-Original-Commit-Position: refs/heads/master@{#371656}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f77609d9f544961f0158fbff93e52c82e4c9dac2