PartitionAlloc: Increase the number of pages per bucket
Currently the number of pages per bucket is limited to 2^16,
and this has caused a bunch of crash reports at partitionBucketFull().

This CL increase the limit to 2^24 instead of decreasing the limit of
|numSystemPagesPerSlotSpan| to 2^8 (It is statically guaranteed that |numSystemPagesPerSlotSpan|
does not exceed 2^8). As a result, this CL doesn't change sizeof(PartitionBucket).

BUG=87772

Review URL: https://codereview.chromium.org/1622553004

Cr-Original-Commit-Position: refs/heads/master@{#371745}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 63f7a4e7c4255204a90569f7ace2d3aeddcea22e