Add metrics for usage of permission features from cross-origin iframes in blink
This CL adds metrics to measure the usage of features that require permissions
in Chrome when they are used in cross-origin iframes only. In particular we only
want to count an iframe if it doesn't have scripting access into the top level
document. A new UseCounter::countCrossOriginIframe function has been added which
will only count this type of usage.

Note that this intends to count cases when a site tries to use an API whether or
not the permission is granted by the user.

BUG=444744

Review URL: https://codereview.chromium.org/1578263006

Cr-Original-Commit-Position: refs/heads/master@{#371756}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe356a2d22a8bf2590ce54d2c9c35cbc8ad0f819