Stop hit testing culled inlines when a match is found
Culled inline hittesting was recently fixed [1] which exposed a
bug in LayoutInline::hitTestCulledInline where hit testing checked
for region intersection but returned a value based on region
containment. This bug was partially fixed in [2] for list-based
tests but a bug still existed for point-based tests.

This patch updates LayoutInline::hitTestCulledInline to check for
hit tests using region intersection, and return a value based on
region intersection (not containment). This was done by fixing a FIXME
to implement HitTestRequest::addNodeToListBasedTestResult for regions.

In addition, this patch changes the return value of
HitTestRequest::addNodeToListBasedTestResult to be an enum which
better explains its behavior.

[1] https://src.chromium.org/viewvc/blink?view=rev&revision=193729
[2] https://crrev.com/b489f7917175e2449c8f6d5874ab3ba0b7f31e88

BUG=579412

Review URL: https://codereview.chromium.org/1630793002

Cr-Original-Commit-Position: refs/heads/master@{#372023}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4dd5f8bbaa8e06c2bca14e20758cd4f76d0a81d4